agents/email_sender/pyproject.toml
[project]
name = "email-sender-agent"
version = "0.1.0"
description = "Sends travel itineraries via Gmail API"
requires-python = ">=3.10,<3.14"
dependencies = [
    "google-adk>=1.27.4,<2.0.0",
    "a2a-sdk==0.3.*",
    "fastapi==0.124.*",
    "uvicorn==0.40.0",
    "google-cloud-logging==3.13.0",
    "opentelemetry-exporter-gcp-trace==1.11.0",
    "httpx>=0.28.1",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
